Restructure remaining packages
