+++ pkg/config/struct_namespace.go
