+++ pkg/config/struct_namespace_test.go
