+++ pkg/discovery/consul.go
@@ -2,7 +2,7 @@ package discovery
 
 import (
 	"github.com/hashicorp/consul/api"
-	"github.com/martin-helmich/prometheus-nginxlog-exporter/config"
+	"github.com/martin-helmich/prometheus-nginxlog-exporter/pkg/config"
 )
 
 // ConsulRegistrator is a helper struct that handles Consul service registration
