+++ pkg/parser/textparser/textparser_test.go
