+++ pkg/syslog/syslog.go
