+++ pkg/tail/struct.go
