+++ pkg/tail/syslog.go
